server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.2",
    "openai>=1.108.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.36.0",
]
